refactor(audio): migrate TX audio pipeline to lock-free SpscRingBuffer
- Add m_faudioRing and m_netAudioRing in TransmitAudioInput for local soundcard/HPSDR and TCI mic inputs
- Push decimated HPSDR mic samples to pushMicAudio in CProtocol1
- Add setTransmitAudioRing in TciServer and push decoded client audio directly to ring buffer
- Drain directly from ring buffers in DataProcessor::fetch_MicData into stack buffer without heap allocations or locks
- Verify lock-free TCI TX audio in tci_server_ws_tests

+++ src/DataEngine/CProtocol1.h
@@ -43,7 +43,7 @@ private:
     int     m_rightSample;
     int     m_micSample;
     float   m_micSample_float;
-    QVector<double> m_hpsdrMicBuffer;
+    QVector<float> m_hpsdrMicBuffer;
     int     m_micDecimCounter = 0;
 };
